bug: implement all feedback - configure eslint to work for react files - fix all eslint errors - use jsx as file extension - change icons [Finishes #150474933]
