+++ client/main.scss
@@ -13,8 +13,16 @@ h1, h2, h3, h4, h5, h6, p {
   padding-top: 10px;
 }
 
+a {
+  font-family: 'Cinzel', serif;
+}
 .nav-anchor {
   font-family: 'Cinzel', serif;
+}
+
+.snav {
+  font-size: 25px !important;
+  color: white !important;
 }
 
 .my-preloader {
@@ -131,6 +139,7 @@ body {
 
 .add-group {
  color: $primary-color;
+ padding-left: 10px;
 }
 
 .modal-overlay {
@@ -144,13 +153,11 @@ body {
 
 #msg-header {
   height: 50px;
-  width: 70%;
-  position: fixed;
   z-index: 999;
 }
 
-.msg-position {
-  margin-top: 50px;
+#user-size {
+  font-size: 12px;
 }
 
 #welcome-header {
@@ -193,6 +200,26 @@ body {
    height: 650px;
  }
 
+.each-group {
+  list-style-type: none;
+}
+
+.groupname-anchor {
+  padding-left: 10px;
+  font-size: 18px;
+}
+
+@media (max-width: 400px) {
+  .groupname-anchor {
+    padding-left: 5px;
+    font-size: 15px;
+  }
+  .each-group {
+    list-style-type: none;
+    margin-left: 5px !important; 
+  }
+}
+
  footer {
    background-color: #333;
    height: 70px;
@@ -212,6 +239,7 @@ body {
    width: 50%;
    height: 30%;
  }
+
  main {
      flex: 1 0 auto;
  }
@@ -342,7 +370,7 @@ body {
 .groupdetails-nav {
   background-color: #76323F;
   width: 320px;
-  height: 550px;
+  height: 565px;
   top: auto;
 }
 
@@ -352,4 +380,30 @@ body {
 
 .group-details {
   color: #1B5E20;
+  padding-left: 10px;
+}
+
+.date-msg-position {
+  font-family: 'Cinzel', serif;
+  display: inline;
+  font-size: 12px;
+}
+
+.msg-content {
+  display: block;
+}
+.date-pad {
+  margin-right: 10px;
+}
+
+#msg-creator {
+  font-family: 'Cinzel', serif;
+}
+
+.wordwrap {  
+   white-space: pre-wrap;      /* Webkit */    
+   white-space: -moz-pre-wrap; /* Firefox */     
+   white-space: -pre-wrap;     /* Opera <7 */    
+   white-space: -o-pre-wrap;   /* Opera 7 */     
+   word-wrap: break-word;      /* IE */ 
 }
