+++ client/src/components/layout/LoginPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import LoginForm from './login/LoginForm.jsx';
+
+const LoginPage = () => {
+  return (
+    <div className="row">
+      <div className="col s6 m9">
+        <LoginForm />
+      </div>
+    </div>
+  );
+};
+
+export default LoginPage;
