+++ client/src/components/layout/Main.jsx
@@ -1,0 +1,24 @@
+import React, { Component } from 'react';
+import NavigationBar from './NavigationBar.jsx';
+
+/**
+ * Main
+ */
+class Main extends Component {
+  /**
+   * render
+   * @return {*} void
+   */
+  render() {
+    return (
+      <div className="landing-background">
+        <NavigationBar />
+        <div className="container-fluid">
+          {this.props.children}
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Main;
