+++ client/src/components/presentation/GroupListSection.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+/**
+ * Home
+ * @return {*} div
+ */
+const GroupListSection = () => {
+  return (
+    <div className="container-fluid">
+      <div className="row" style={{ marginBottom: 0 }}>
+        <div className="col s12 m12" style={{ paddingLeft: 0 }}>
+          <div className="container-fluid" />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default GroupListSection;
