+++ client/src/components/presentation/PreLoader.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import { Link } from 'react-router';
+
+/**
+ * Home
+ * @return {*} div
+ */
+const PreLoader = () => {
+  return (
+    <div className="container">
+      <div className="col s3 offset-s4 my-preloader">
+        <div className="valign-wrapper preloader-wrapper active">
+          <div className="spinner-layer spinner-red-only">
+            <div className="circle-clipper left">
+              <div className="circle" />
+            </div>
+            <div className="gap-patch">
+              <div className="circle" />
+            </div>
+            <div className="circle-clipper right">
+              <div className="circle" />
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default PreLoader;
