+++ client/src/components/presentation/index.js
@@ -1,13 +1,21 @@
-import CreateMessage from './CreateMessage';
-import Group from './Group';
-import Message from './Message';
-import AddUserModal from './AddUserModal';
-import SideNav from './SideNav';
+import CreateMessage from './CreateMessage.jsx';
+import Group from './Group.jsx';
+import Message from './Message.jsx';
+import AddUserModal from './AddUserModal.jsx';
+import SideNav from './SideNav.jsx';
+import GroupListSection from './GroupListSection.jsx';
+import GroupHeader from './GroupHeader.jsx';
+import SearchButton from './SearchButton.jsx';
+import PreLoader from './PreLoader.jsx';
 
 export {
   CreateMessage,
   Group,
   Message,
   AddUserModal,
-  SideNav
+  SideNav,
+  GroupListSection,
+  GroupHeader,
+  SearchButton,
+  PreLoader
 };
