+++ client/src/reducers/groupReducer.js
@@ -1,11 +1,9 @@
-import { GROUPS_RECEIVED, GROUPS_NOT_RECEIVED, GROUP_CREATE, GROUP_SELECTED, APPLICATION_STATE, GROUP_MESSAGES, ADD_USER, SEND_MESSAGE } from '../actions/types';
+import { GROUPS_RECEIVED, GROUPS_NOT_RECEIVED, GROUP_CREATE, GROUP_SELECTED, APPLICATION_STATE, ADD_USER } from '../actions/types';
 
 const initialState = {
   groupList: [],
   selectedGroup: null,
-  groupMessages: [],
   appStatus: 'ready',
-  list: {}
 };
 /**
  * Group reducer takes the state and action and then returns the state
@@ -16,40 +14,32 @@ const initialState = {
  */
 export default (state = initialState, action = {}) => {
   const updated = Object.assign({}, state);
-  const updatedList = Object.assign([], updated['groupList']);
+  const updatedList = Object.assign([], updated.groupList);
   switch (action.type) {
     case GROUPS_RECEIVED:
-      updated['groupList'] = action.groups;
-      updated['appStatus'] = 'ready';
+      updated.groupList = action.groups;
+      updated.appStatus = 'ready';
       return updated;
 
     case GROUPS_NOT_RECEIVED:
-      updated['appStatus'] = 'no groups';
+      updated.appStatus = 'no groups';
       return updated;
 
     case GROUP_CREATE:
       updatedList.push(action.group);
-      updated['groupList'] = updatedList;
+      updated.groupList = updatedList;
       return updated;
 
     case GROUP_SELECTED:
-      updated['selectedGroup'] = action.selectedGroup;
+      updated.selectedGroup = action.selectedGroup;
       return updated;
 
     case APPLICATION_STATE:
-      updated['appStatus'] = action.status;
-      return updated;
-
-    case GROUP_MESSAGES:
-      updated['groupMessages'] = action.messages;
+      updated.appStatus = action.status;
       return updated;
 
     case ADD_USER:
-      updated['addUser'] = action.addUser;
-      return updated;
-
-    case SEND_MESSAGE:
-      updated['list'] = action.message;
+      updated.addUser = action.addUser;
       return updated;
 
     default:
