+++ client/src/reducers/userReducer.js
@@ -8,7 +8,7 @@ export default (state = initialState, action = {}) => {
   const updated = Object.assign({}, state);
   switch (action.type) {
     case GET_ALL_USERS:
-      updated['users'] = action.users;
+      updated.users = action.users;
       return updated;
 
     default: return state;
