--- client/src/utils/index.js
@@ -1,5 +1,0 @@
-import APIManager from './APIManager';
-
-export {
-  APIManager
-};
