+++ package.json
@@ -51,6 +51,7 @@
     "express-favicon": "^2.0.0",
     "express-session": "^1.15.3",
     "file-loader": "^0.11.2",
+    "history": "^4.6.3",
     "identity-obj-proxy": "^3.0.0",
     "image-webpack-loader": "^3.3.1",
     "jest": "^20.0.4",
